Set user after successfull authentication
