+++ src/redux/reducers.js
@@ -1,5 +1,5 @@
-import auth from './auth/reducer';
+import user from './user/reducer';
 
 export default {
-    auth,
+    user,
 };
