+++ src/redux/sagas.js
@@ -1,8 +1,8 @@
 import { all } from 'redux-saga/effects';
-import authSagas from './auth/saga';
+// import authSagas from './user/saga';
 
 export default function* rootSaga() {
     yield all([
-        authSagas(),
+        // authSagas(),
     ]);
 }
